Renaming: from synthetic to isa_regdeps.
