+++ core/CMakeLists.txt
@@ -279,8 +279,8 @@ set(DECODER_SRCS
   ir/${ARCH_NAME}/decode.c
   ir/encode_shared.c
   ir/${ARCH_NAME}/encode.c
-  ir/synthetic/encode.c
-  ir/synthetic/decode.c
+  ir/isa_regdeps/encode.c
+  ir/isa_regdeps/decode.c
   ir/disassemble_shared.c
   ir/${ARCH_NAME}/disassemble.c
   ir/ir_utils_shared.c
