+++ core/ir/arm/decode.c
@@ -31,7 +31,7 @@
  */
 
 #include "../globals.h"
-#include "../synthetic/decode.h"
+#include "../isa_regdeps/decode.h"
 #include "encode_api.h"
 #include "instr.h"
 #include "decode.h"
