+++ core/ir/isa_regdeps/decode.h
@@ -30,12 +30,12 @@
  * DAMAGE.
  */
 
-#ifndef _SYNTHETIC_DECODE_H_
-#define _SYNTHETIC_DECODE_H_ 1
+#ifndef _REGDEPS_DECODE_H_
+#define _REGDEPS_DECODE_H_ 1
 
 #include "../globals.h"
 
 byte *
 decode_from_synth(dcontext_t *dcontext, byte *encoded_instr, instr_t *instr);
 
-#endif /* _SYNTHETIC_DECODE_H_ */
+#endif /* _REGDEPS_DECODE_H_ */
