+++ core/ir/isa_regdeps/encode.h
@@ -30,12 +30,12 @@
  * DAMAGE.
  */
 
-#ifndef _SYNTHETIC_ENCODE_H_
-#define _SYNTHETIC_ENCODE_H_ 1
+#ifndef _REGDEPS_ENCODE_H_
+#define _REGDEPS_ENCODE_H_ 1
 
 #include "../globals.h"
 
 byte *
 encode_to_synth(dcontext_t *dcontext, instr_t *instr, byte *encoded_instr);
 
-#endif /* _SYNTHETIC_ENCODE_H_ */
+#endif /* _REGDEPS_ENCODE_H_ */
