+++ suite/tests/api/ir_regdeps.expect
